Added import and did some reshuffling of functions
