+++ ubelt/__init__.py
@@ -32,6 +32,7 @@ GLOBAL_MODULES = [
     'util_cache',
     'util_colors',
     'util_const',
+    'util_cmd',
     'util_decor',
     'util_dict',
     'util_download',
@@ -39,6 +40,7 @@ GLOBAL_MODULES = [
     'util_format',
     'util_io',
     'util_list',
+    'util_import',
     'util_mixins',
     'util_path',
     'util_platform',
@@ -62,10 +64,12 @@ else:  # pragma: nobranch
 
 if _DOELSE:  # pragma: nobranch
     # <AUTOGEN_INIT>
+
     from ubelt import util_arg
     from ubelt import util_cache
     from ubelt import util_colors
     from ubelt import util_const
+    from ubelt import util_cmd
     from ubelt import util_decor
     from ubelt import util_dict
     from ubelt import util_download
@@ -73,6 +77,7 @@ if _DOELSE:  # pragma: nobranch
     from ubelt import util_format
     from ubelt import util_io
     from ubelt import util_list
+    from ubelt import util_import
     from ubelt import util_mixins
     from ubelt import util_path
     from ubelt import util_platform
@@ -85,31 +90,34 @@ if _DOELSE:  # pragma: nobranch
     from ubelt.util_cache import (Cacher,)
     from ubelt.util_colors import (color_text, highlight_code,)
     from ubelt.util_const import (NoParam,)
+    from ubelt.util_cmd import (cmd,)
     from ubelt.util_decor import (memoize,)
-    from ubelt.util_dict import (AutoDict, AutoOrderedDict, ddict, dict_hist,
-                                 dict_subset, dict_take, find_duplicates,
-                                 group_items, invert_dict, map_keys, map_vals,
-                                 odict,)
-    from ubelt.util_download import (download, grabdata,)
+    from ubelt.util_dict import (AutoDict, AutoOrderedDict, OrderedDict, ddict,
+                                 defaultdict, dict_hist, dict_subset, dict_take,
+                                 find_duplicates, group_items, invert_dict,
+                                 map_keys, map_vals, odict,)
+    from ubelt.util_download import (download, grabdata, urlopen, urlparse,)
     from ubelt.util_func import (identity,)
     from ubelt.util_format import (repr2,)
-    from ubelt.util_io import (delete, readfrom, touch, writeto,)
+    from ubelt.util_io import (readfrom, writeto, touch, delete,)
     from ubelt.util_list import (argsort, boolmask, chunks, compress, flatten,
                                  iterable, take, unique, unique_flags,)
+    from ubelt.util_import import (PythonPathContext, import_module_from_name,
+                                   import_module_from_path,)
     from ubelt.util_mixins import (NiceRepr,)
-    from ubelt.util_path import (augpath, compressuser, truepath, userhome,)
-    from ubelt.util_platform import (DARWIN, LINUX, POSIX, PY2, PY3, WIN32, cmd,
+    from ubelt.util_path import (abspath, augpath, compressuser, ensuredir,
+                                 expandvars, split, truepath, userhome,)
+    from ubelt.util_platform import (DARWIN, LINUX, POSIX, PY2, PY3, WIN32,
                                      editfile, ensure_app_cache_dir,
-                                     ensure_app_resource_dir, ensuredir,
-                                     get_app_cache_dir, get_app_resource_dir,
-                                     platform_cache_dir, platform_resource_dir,
-                                     startfile,)
-    from ubelt.util_str import (CaptureStdout, codeblock, ensure_unicode, hzcat,
-                                indent,)
+                                     ensure_app_resource_dir, get_app_cache_dir,
+                                     get_app_resource_dir, platform_cache_dir,
+                                     platform_resource_dir, startfile,)
+    from ubelt.util_str import (CaptureStdout, cStringIO, codeblock,
+                                ensure_unicode, hzcat, indent,)
     from ubelt.util_stress import (find_nth_prime,)
     from ubelt.util_time import (Timer, Timerit, timestamp,)
     from ubelt.progiter import (ProgIter,)
-
+    
     # </AUTOGEN_INIT>
 
 del _DOELSE
